Scrapes from a more reliable URL. Renamed file and added to main.py.
