+++ extractors/main.py
@@ -2,6 +2,7 @@ from extractors.blockextractor import BlockExtractor
 from extractors.characterfactory import CharacterFactory
 from extractors.emojiextractor import EmojiExtractor
 from extractors.mathcollectionextractor import MathExtractor
+from extractors.nerdextractor import NerdExtractor
 
 if __name__ == "__main__":
     character_factory = CharacterFactory()
@@ -9,3 +10,4 @@ if __name__ == "__main__":
     EmojiExtractor().extract()
     BlockExtractor(character_factory).extract()
     MathExtractor(character_factory).extract()
+    NerdExtractor().extract()
